--- Packages/Clustering/test.yaml
@@ -1,4 +1,0 @@
-clusters: !!set
-  ? 1
-  ? 2
-  ? 3
